requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf2
python-dotenv
google-generativeai